+++ addons/pinniped/tanzu-auth-controller/controllers/utils.go
