+++ addons/pinniped/tanzu-auth-controller/hack/test.sh
@@ -8,11 +8,11 @@ set -euo pipefail
 MY_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 TF_ROOT="${MY_DIR}/../../../.."
 
-# Always run from config-controller directory for reproducibility
+# Always run from tanzu-auth-controller directory for reproducibility
 cd "${MY_DIR}/.."
 
-# Install golangci-lint
-make golangci-lint -C "${TF_ROOT}/hack/tools"
+# Install kubebuilder
+make kubebuilder -C "${TF_ROOT}/hack/tools"
 
-# Run linter
-"${TF_ROOT}/hack/tools/bin/golangci-lint" run -v
+# Run tests
+KUBEBUILDER_ASSETS="${TF_ROOT}/hack/tools/bin/kubebuilder/bin" go test ./... -coverprofile cover.out -v 2
