+++ addons/pinniped/tanzu-auth-controller/hack/ytt/default-package-secret.yaml
