+++ addons/pinniped/tanzu-auth-controller/hack/ytt/schema.yaml
